==> Code/Data_process/1_Get_D_P_M.ipynb <==
import numpy as np
import dill
import pandas as pd

import datetime as dt
from tqdm import tqdm
from collections import defaultdict, Counter

import warnings
# 禁用所有警告
warnings.filterwarnings("ignore")

df = pd.read_csv('output/III_patients.csv')
print(df.shape,len(df.HADM_ID.unique()),len(df.ICUSTAY_ID.unique()))
print('DIEINHOSPITAL:',df['DIEINHOSPITAL'].value_counts())

date_columns = ['ADMITTIME', 'DISCHTIME', 'INTIME', 'OUTTIME','DOD']
df[date_columns] = df[date_columns].apply(pd.to_datetime)

## Diagnosis

#note:use MIMIC-IV D_ICD_DIAGNOSES
D_ICD_DIAGNOSES = pd.read_csv('input/d_icd_diagnoses.csv.gz')
D_ICD_DIAGNOSES.columns = D_ICD_DIAGNOSES.columns.str.upper()
D_ICD_DIAGNOSES.columns = ['ICD_CODE','ICD_VERSION','ICD_TEXT']

diag_9_10 = pd.read_csv('input/0_diag_9_10.csv')
diag_9_10.columns = ['ICD_CODE', 'ROOT']
diag_9_10 = diag_9_10[~diag_9_10.duplicated(subset=['ICD_CODE'], keep=False)]
diag_9_10['ROOT'] = diag_9_10['ROOT'].fillna(diag_9_10['ICD_CODE'])
print(diag_9_10.shape,len(diag_9_10.ICD_CODE.unique()))

#included diagnose id
icds = ['Z85', 'J98', 'I25', 'F10', 'J69', 'J45', 'J18', 'E89', 'D69', 'I95', 'I10', 'E66', 'E87', 'R78', 'A40', 'N17', 'F41', 'N18', 'G47', 'E11', 'R09', 'I47', 'K70', '9971', 'Z95', 'I21', 'I12', 'F32', 'I34', 'T78', 'K22', 'I50', 'J44', 'D64', 'R00', 'E78', 'N39', 'K92']
len(icds)

# MIMIC-III
III_diagnoses_raw = pd.read_csv('input/DIAGNOSES_ICD.csv.gz')

# Set ICD version to 9 for MIMIC-III data
III_diagnoses_raw['ICD_VERSION'] = 9

# Print shape and unique ICD9 codes
III_diagnoses_raw = III_diagnoses_raw[['SUBJECT_ID', 'HADM_ID', 'ICD9_CODE', 'ICD_VERSION']]
III_diagnoses_raw.columns = ['SUBJECT_ID', 'HADM_ID', 'ICD_CODE', 'ICD_VERSION']
print(III_diagnoses_raw.shape, len(III_diagnoses_raw.ICD_CODE.unique()))

# Merge raw diagnoses data with the ICD-9 to ICD-10 conversion table
III_diagnoses = pd.merge(
    III_diagnoses_raw,
    diag_9_10,
    on='ICD_CODE',
    how='left'
)

# Fill missing ROOT values with ICD_CODE and remove duplicates
III_diagnoses['ROOT'] = III_diagnoses['ROOT'].fillna(III_diagnoses['ICD_CODE'])
III_diagnoses = III_diagnoses.drop_duplicates(subset=['HADM_ID', 'ICD_CODE', 'ICD_VERSION', 'ROOT'], keep='first')

# Print updated shape, unique ICD codes, and ROOT values
III_diagnoses = III_diagnoses[III_diagnoses['ROOT'].isin(icds)]
print(III_diagnoses.shape, len(III_diagnoses.ICD_CODE.unique()), len(III_diagnoses.ROOT.unique()))
print(list(III_diagnoses.ROOT.unique()))

# Merge with ICD text descriptions for ICD-9 codes
III_diagnoses = pd.merge(
    III_diagnoses,
    D_ICD_DIAGNOSES[D_ICD_DIAGNOSES.ICD_VERSION == 9][['ICD_CODE', 'ICD_TEXT']],
    on='ICD_CODE',
    how='left'
)

# Group by subject and admission ID to list ICD codes per ROOT category
III_diagnoses_result = III_diagnoses.groupby(['SUBJECT_ID', 'HADM_ID', 'ROOT'])['ICD_CODE'].apply(list).reset_index()
# Pivot the grouped data to create a matrix for HADM_ID and ROOT categories
III_diagnoses_result = III_diagnoses_result.pivot(index='HADM_ID', columns='ROOT', values='ICD_CODE')
# Reset the index after pivot
III_diagnoses_result = III_diagnoses_result.reset_index()
# Display the first few rows of the final result
print(III_diagnoses_result.head())

#Check nan
print(III_diagnoses[III_diagnoses.ICD_TEXT.isna()].shape)

III_diagnoses_result.loc[:, III_diagnoses_result.columns != 'HADM_ID'] = III_diagnoses_result.loc[:, III_diagnoses_result.columns != 'HADM_ID'].notna().astype(int)
print(III_diagnoses_result.shape,len(III_diagnoses_result.HADM_ID.unique()))

III_diagnoses_result = III_diagnoses_result[III_diagnoses_result.HADM_ID.isin(df.HADM_ID)]
print(III_diagnoses_result.shape,len(III_diagnoses_result.HADM_ID.unique()))

III_diagnoses_result.columns = ['Diag_' + col if (col != 'HADM_ID')&(col != 'ICUSTAY_ID') else col for col in III_diagnoses_result.columns]
III_diagnoses_result.head()

## Procedures

#note:use MIMIC-IV d_icd_procedures
d_icd_procedures = pd.read_csv('input/d_icd_procedures.csv.gz')
d_icd_procedures.columns = ['ICD_CODE', 'ICD_VERSION', 'ICD_TEXT']
print(d_icd_procedures.ICD_VERSION.value_counts())
d_icd_procedures.head(2)

#included Procedures id
icds = [9604, 3722, 3723, 225966, 8856, 224270, 3995, 224385, 3615, 225402, 4513, 8872, 9390, 9904, 3891, 3893, 225401, 3897, 9915, 224264, 966, 9671, 9672, 221217, 3961, 221214, 225792, 225454, 225752, 227194, 3324]
len(icds)

III_p_raw = pd.read_csv('input/PROCEDURES_ICD.csv.gz')
III_p_raw = III_p_raw[['SUBJECT_ID', 'HADM_ID','ICD9_CODE']]
III_p_raw.columns = ['SUBJECT_ID', 'HADM_ID','ICD_CODE']
III_p_raw = III_p_raw.drop_duplicates(keep='first')
print('III_p_raw',III_p_raw.shape,len(III_p_raw.HADM_ID.unique()))

III_p_raw_result = III_p_raw[III_p_raw.ICD_CODE.isin(icds)][['HADM_ID','ICD_CODE']].pivot_table(index='HADM_ID', columns='ICD_CODE', aggfunc='size', fill_value=np.nan)
III_p_raw_result = III_p_raw_result.reset_index()

III_p_raw_result.loc[:, III_p_raw_result.columns != 'HADM_ID'] = III_p_raw_result.loc[:, III_p_raw_result.columns != 'HADM_ID'].notna().astype(int)
print('III_p_raw_result',III_p_raw_result.shape,len(III_p_raw_result.HADM_ID.unique()))

III_p_raw_result = III_p_raw_result[III_p_raw_result.HADM_ID.isin(df.HADM_ID)]
print('III_p_raw_result',III_p_raw_result.shape,len(III_p_raw_result.HADM_ID.unique()))

event_p = set(icds)-set(III_p_raw_result.columns)
print('Add procedure from event:',len(event_p))
proc = pd.read_csv('input/PROCEDUREEVENTS_MV.csv.gz')
print(proc.shape,len(proc.ICUSTAY_ID.unique()))
proc = proc[['ICUSTAY_ID', 'STARTTIME', 'ENDTIME','ITEMID', 'VALUE','ORDERCATEGORYNAME']]
proc = proc[proc['ICUSTAY_ID'].isin(df['ICUSTAY_ID'].unique())]
print(proc.shape,len(proc.ICUSTAY_ID.unique()))
proc = proc[proc['ITEMID'].isin(event_p)]
proc = proc.dropna(subset=['VALUE'])
print(proc.shape,len(proc.ICUSTAY_ID.unique()))

proc['STARTTIME'] = pd.to_datetime(proc['STARTTIME'])
proc['ENDTIME'] = pd.to_datetime(proc['ENDTIME'])
proc = pd.merge(proc,df[['ICUSTAY_ID', 'INTIME', 'OUTTIME']],on='ICUSTAY_ID',how='left')
within_criteria = proc['STARTTIME'].between(proc['INTIME'],proc['OUTTIME'])
within = proc[within_criteria]
print(within.shape)

within = within[['ICUSTAY_ID','ITEMID']]
within = within.drop_duplicates(keep='first')
print(within.shape)

within_result = within.pivot_table(index='ICUSTAY_ID', columns='ITEMID', aggfunc='size', fill_value=np.nan)
within_result = within_result.reset_index()

within_result.loc[:, within_result.columns != 'ICUSTAY_ID'] = within_result.loc[:, within_result.columns != 'ICUSTAY_ID'].notna().astype(int)
print('within_result',within_result.shape,len(within_result.ICUSTAY_ID.unique()))

within_result = within_result[within_result.ICUSTAY_ID.isin(df.ICUSTAY_ID)]
print('within_result',within_result.shape,len(within_result.ICUSTAY_ID.unique()))

final_p = df[['HADM_ID','ICUSTAY_ID']]
final_p = pd.merge(final_p,III_p_raw_result,on='HADM_ID',how='left')
final_p = pd.merge(final_p,within_result,on='ICUSTAY_ID',how='left')
print(final_p.shape,len(final_p.HADM_ID.unique()),len(final_p.ICUSTAY_ID.unique()))

final_p = final_p[~final_p[final_p.columns.difference(['HADM_ID', 'ICUSTAY_ID'])].isna().all(axis=1)]
final_p = final_p.fillna(0)
print(final_p.shape,len(final_p.HADM_ID.unique()),len(final_p.ICUSTAY_ID.unique()))

final_p.columns = final_p.columns.astype(str)
final_p.columns = ['Pro_' + col if (col != 'HADM_ID')&(col != 'ICUSTAY_ID') else col for col in final_p.columns]
final_p = final_p.drop('HADM_ID', axis=1)
final_p.head()

## Drugs

druginfo = pd.read_csv('input/drugbank_drugs_info.csv')

rxnorm2atc4 = pd.read_csv('input/RXCUI2atc4.csv')
rxnorm2atc4 = rxnorm2atc4.drop(columns=['YEAR', 'MONTH', 'NDC'])
rxnorm2atc4.drop_duplicates(subset=['RXCUI'], inplace=True)

with open('input/rxnorm2RXCUI.txt', 'r') as f:
    rxnorm2RXCUI = eval(f.read())
    
#included Drugs id
drug_ids = ['A02AC', 'A02BA', 'A02BC', 'A04AA', 'A06AB', 'A06AD', 'A07AA', 'B01AA', 'B01AC', 'C01BD', 'C03CA', 'C07AB', 'C09AA', 'J01MA', 'N02BE', 'N05BA', 'R01AX']
print(len(drug_ids))

def ATC3toDrug(med_pd):
    atc3toDrugDict = {}
    for atc3, drugname in med_pd[['ATC3', 'DRUG']].values:
        if atc3 in atc3toDrugDict:
            atc3toDrugDict[atc3].add(drugname)  
        else:
            atc3toDrugDict[atc3] = {drugname}  

    return atc3toDrugDict

def atc3toSMILES(ATC3toDrugDict, druginfo):
    drug2smiles = {}
    atc3tosmiles = {}
    for drugname, smiles in druginfo[['name', 'moldb_smiles']].values:
        if type(smiles) == type('a'):
            drug2smiles[drugname] = smiles
    for atc3, drug in ATC3toDrugDict.items():
        temp = []
        for d in drug:
            try:
                temp.append(drug2smiles[d])
            except:
                pass
        if len(temp) > 0:
            atc3tosmiles[atc3] = temp[:3]
    
    return atc3tosmiles

III_PRESCRIPTIONS = pd.read_csv('input/PRESCRIPTIONS.csv.gz', dtype={'NDC':'category'})
print(III_PRESCRIPTIONS.shape)
III_PRESCRIPTIONS = III_PRESCRIPTIONS[III_PRESCRIPTIONS.HADM_ID.isin(df.HADM_ID)]
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS.STARTDATE = pd.to_datetime(III_PRESCRIPTIONS.STARTDATE)
III_PRESCRIPTIONS.ENDDATE = pd.to_datetime(III_PRESCRIPTIONS.ENDDATE)

III_PRESCRIPTIONS_notna = III_PRESCRIPTIONS[~III_PRESCRIPTIONS['ICUSTAY_ID'].isna()]
III_PRESCRIPTIONS_notna = III_PRESCRIPTIONS_notna[III_PRESCRIPTIONS_notna.ICUSTAY_ID.isin(df.ICUSTAY_ID)]
print(III_PRESCRIPTIONS_notna.shape)

III_PRESCRIPTIONS_isna = III_PRESCRIPTIONS[III_PRESCRIPTIONS['ICUSTAY_ID'].isna()]
III_PRESCRIPTIONS_isna = III_PRESCRIPTIONS_isna.drop(['ICUSTAY_ID'], axis=1)
III_PRESCRIPTIONS_isna = pd.merge(III_PRESCRIPTIONS_isna, df[['HADM_ID','ICUSTAY_ID']], on='HADM_ID', how='left')
III_PRESCRIPTIONS_isna = III_PRESCRIPTIONS_isna[III_PRESCRIPTIONS_isna.ICUSTAY_ID.isin(df.ICUSTAY_ID)]
print(III_PRESCRIPTIONS_isna.shape)

III_PRESCRIPTIONS = pd.concat([III_PRESCRIPTIONS_notna,III_PRESCRIPTIONS_isna])
III_PRESCRIPTIONS = III_PRESCRIPTIONS[III_PRESCRIPTIONS.ICUSTAY_ID.isin(df.ICUSTAY_ID)]
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS = pd.merge(III_PRESCRIPTIONS, df.drop(['HADM_ID'], axis=1), on='ICUSTAY_ID', how='left')
III_PRESCRIPTIONS['STARTDATE_IN_RANGE'] = (III_PRESCRIPTIONS['INTIME'].dt.date <= III_PRESCRIPTIONS['STARTDATE']) & (III_PRESCRIPTIONS['STARTDATE'] <= III_PRESCRIPTIONS['OUTTIME'].dt.date)
III_PRESCRIPTIONS['STARTDATE_afterICU'] = (III_PRESCRIPTIONS['OUTTIME'].dt.date <= III_PRESCRIPTIONS['STARTDATE']) & (III_PRESCRIPTIONS['STARTDATE'] <= III_PRESCRIPTIONS['DISCHTIME'].dt.date)

III_PRESCRIPTIONS = III_PRESCRIPTIONS[III_PRESCRIPTIONS.STARTDATE_afterICU==True]
III_PRESCRIPTIONS = III_PRESCRIPTIONS[['HADM_ID','ICUSTAY_ID','DRUG','NDC']]
print(III_PRESCRIPTIONS.shape)

# Step 1: check nan NDC
drugs_with_non_empty_ndc = III_PRESCRIPTIONS.dropna(subset=['NDC']).groupby('DRUG').filter(lambda x: x['NDC'].notna().any())
III_PRESCRIPTIONS = III_PRESCRIPTIONS[III_PRESCRIPTIONS['DRUG'].isin(drugs_with_non_empty_ndc['DRUG'])]

# Step 2: group DRUG
most_common_ndc = III_PRESCRIPTIONS.dropna(subset=['NDC']).groupby('DRUG')['NDC'].agg(lambda x: x.mode()[0])

# Step 3: fillna
III_PRESCRIPTIONS['NDC'] = III_PRESCRIPTIONS.apply(lambda row: most_common_ndc[row['DRUG']] if pd.isna(row['NDC']) else row['NDC'], axis=1)

III_PRESCRIPTIONS = III_PRESCRIPTIONS.reset_index(drop=True)
III_PRESCRIPTIONS = III_PRESCRIPTIONS[~(III_PRESCRIPTIONS.NDC=='0')]
III_PRESCRIPTIONS = III_PRESCRIPTIONS.drop_duplicates(keep='first')
III_PRESCRIPTIONS = III_PRESCRIPTIONS.reset_index(drop=True)
print(III_PRESCRIPTIONS.shape,len(III_PRESCRIPTIONS.HADM_ID.unique()),len(III_PRESCRIPTIONS.ICUSTAY_ID.unique()))

III_PRESCRIPTIONS = III_PRESCRIPTIONS[['ICUSTAY_ID','DRUG','NDC']]
III_PRESCRIPTIONS = III_PRESCRIPTIONS.drop_duplicates(keep='first')
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS['RXCUI'] = III_PRESCRIPTIONS['NDC'].map(rxnorm2RXCUI)
III_PRESCRIPTIONS.dropna(inplace=True)
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS = III_PRESCRIPTIONS[~(III_PRESCRIPTIONS.RXCUI == '')]
III_PRESCRIPTIONS['RXCUI'] = III_PRESCRIPTIONS['RXCUI'].astype('int64')
III_PRESCRIPTIONS = III_PRESCRIPTIONS.reset_index(drop=True)
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS = III_PRESCRIPTIONS.merge(rxnorm2atc4, on=['RXCUI'])
III_PRESCRIPTIONS = III_PRESCRIPTIONS.rename(columns={'ATC4': 'ATC3'})
print(III_PRESCRIPTIONS.shape)

III_PRESCRIPTIONS = III_PRESCRIPTIONS[III_PRESCRIPTIONS.ATC3.isin(drug_ids)]
print(III_PRESCRIPTIONS.shape,len(III_PRESCRIPTIONS.ATC3.unique()),len(III_PRESCRIPTIONS.NDC.unique()))

III_atc3toDrug = ATC3toDrug(III_PRESCRIPTIONS)
print(len(III_atc3toDrug))

III_SMIL = atc3toSMILES(III_atc3toDrug,druginfo)
print(len(III_SMIL))

III_med_pd = III_PRESCRIPTIONS[III_PRESCRIPTIONS.ATC3.isin(III_SMIL.keys())]
print(III_med_pd.shape,len(III_med_pd.ATC3.unique()),len(III_med_pd.NDC.unique()))
III_med_pd['SMILES'] = III_med_pd['ATC3'].map(lambda x: '\t'.join(III_SMIL[x]))
print(III_med_pd.shape,len(III_med_pd.ATC3.unique()),len(III_med_pd.NDC.unique()))

III_m = III_med_pd[['ICUSTAY_ID','ATC3']]
III_m = III_m.drop_duplicates(keep='first')

III_m = III_m.pivot_table(index='ICUSTAY_ID', columns='ATC3', aggfunc='size', fill_value=np.nan)
III_m = III_m.reset_index()

III_m.loc[:, III_m.columns != 'ICUSTAY_ID'] = III_m.loc[:, III_m.columns != 'ICUSTAY_ID'].notna().astype(int)
print(III_m.shape)

III_m.columns = ['Drug_' + col if (col != 'HADM_ID')&(col != 'ICUSTAY_ID') else col for col in III_m.columns]
III_m.head()

## Merge

print(III_diagnoses_result.columns)
print(final_p.columns)
print(III_m.columns)

Final_III = pd.merge(df,III_diagnoses_result,on='HADM_ID',how='left')
Final_III = pd.merge(Final_III,final_p,on='ICUSTAY_ID',how='left')
Final_III = pd.merge(Final_III,III_m,on='ICUSTAY_ID',how='left')
print(Final_III.shape,len(Final_III.HADM_ID.unique()),len(Final_III.ICUSTAY_ID.unique()))

Final_III = Final_III[~Final_III[Final_III.columns.difference(df.columns)].isna().all(axis=1)]
Final_III = Final_III.fillna(0)
print(Final_III.shape,len(Final_III.HADM_ID.unique()),len(Final_III.ICUSTAY_ID.unique()))
Final_III['DIEINHOSPITAL'].value_counts()

Final_III.head()

Final_III.to_csv('output/Final_III_0.csv',index=False)
